JDHJMerge: updated Gui: weitere freeze bugs fixed
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@5992 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/optional/hjsplit/JDHJSplit.java
@@ -269,13 +269,16 @@ public class JDHJSplit extends PluginOptional implements ControlListener {
 
                     if (output.exists()) output.delete();
                 }
-                join.run();
-                if (getPluginConfig().getBooleanProperty(CONFIG_KEY_REMOVE_MERGED, false)) {
-                    ArrayList<File> list = getFileList(new File(link.getFileOutput()));
-                    for (File f : list) {
-                        f.delete();
-                        f.deleteOnExit();
+                try {
+                    join.run();
+                    if (join.wasSuccessfull() && getPluginConfig().getBooleanProperty(CONFIG_KEY_REMOVE_MERGED, false)) {
+                        ArrayList<File> list = getFileList(new File(link.getFileOutput()));
+                        for (File f : list) {
+                            f.delete();
+                            f.deleteOnExit();
+                        }
                     }
+                } catch (Exception e) {
                 }
                 progress.finalize();
                 JDUtilities.getController().fireControlEvent(new ControlEvent(this, ControlEvent.CONTROL_ON_FILEOUTPUT, new File[] { output }));
